Add EffVer to release notes (#34)
As is, this PR results in generating release notes as so...

# napari 0.6.5

We’re happy to announce the release of napari 0.6.5!

napari is a fast, interactive, multi-dimensional image viewer for
Python. It’s designed for exploring, annotating, and analyzing
multi-dimensional images. It’s built on Qt (for the GUI), VisPy (for
performant GPU-based rendering), and the scientific Python stack (NumPy,
SciPy, and friends).

For more information, examples, and documentation, please visit our
website: https://napari.org/

napari follows [EffVer (Intended Effort
Versioning)](https://effver.org/); this is a **Micro** release
containing awesome new features that are expected to be adoptable with
no additional effort.

## Highlights

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>